Moved termination test to be run before others.
